# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Xoycoin Exchange",
    "description": "Métodos de la API de Xoycoin Exchange",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-xoycoin.xellcoders.com",
      "description": "Xoycoin Exchange API"
    }
  ],
...
  "paths": {
    "/access": {
      "get": {
        "tags": [
          "access"
        ],
        "summary": "Get user data",
        "description": "Obtiene los datos del usuario asociado al token de acceso.",
        "responses": {
          "200": {
            "description": "Access object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Access"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "email": "exchange@domain.com",
                  "status": "LOCKED"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Access not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "access"
        ],
        "summary": "Refresh access session",
        "description": "Resfresca la vigencia del token de acceso.",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "access"
        ],
        "summary": "Authentication",
        "description": "Autenticación",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/access_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Access object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Access"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "email": "exchange@domain.com",
                  "status": "LOCKED",
                  "token": "6dID6EZCZXXzZCBPJRchmrEP2uWb7TpI1J1rRHpvFh15mc57DY5aI41gEJ9pNRNv"
                }
              }
            }
          },
          "500": {
            "description": "ERROR: Access not found"
          }
        }
      },
      "delete": {
        "tags": [
          "access"
        ],
        "summary": "Delete access session",
        "description": "Elimina la sesión del token de acceso.",
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/access/{accessId}": {
      "get": {
        "tags": [
          "access"
        ],
        "summary": "Get access data by id",
        "description": "Obtiene los datos del usuario mediante el id proporcionado.",
        "parameters": [
          {
            "name": "accessId",
            "in": "path",
            "description": "Id of the access to return",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Access"
                },
                "example": {
                  "id": "a860b92d-093d-4fe7-a7d5-e810cb7fcd3a",
                  "email": "exchange@domain.com",
                  "status": "LOCKED"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Access not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/access/activate": {
      "put": {
        "tags": [
          "access"
        ],
        "summary": "Activate access",
        "description": "Activa un usaurio.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/access_activate_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/access/inactivate": {
      "put": {
        "tags": [
          "access"
        ],
        "summary": "Inactivate access",
        "description": "Inactiva un usuario.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/access_inactivate_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/access/lock": {
      "put": {
        "tags": [
          "access"
        ],
        "summary": "Lock access",
        "description": "Bloquea un usuario.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/access_lock_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "example": {
                  "result": true
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Access": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "INACTIVE",
              "ACTIVE",
              "LOCKED",
              "DISABLED"
            ]
          }
        }
      },
...
      "access_body": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "access_activate_body": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "access_inactivate_body": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "access_lock_body": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}